refactor: 수정된 디렉토리명 반영
